requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
async-timeout==5.0.1
autopep8==2.3.1
billiard==4.2.1
blinker==1.9.0
Brotli==1.1.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
cryptography==43.0.3
defusedxml==0.8.0rc2
distlib==0.3.9
distro==1.9.0
dj-database-url==2.3.0
Django==4.2.16
django-cors-headers==4.6.0
django-debug-toolbar==4.4.6
django-filter==24.3
django-redis==5.4.0
django-silk==5.3.1
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
docopt==0.6.2
drf-nested-routers==0.94.1
exceptiongroup==1.2.2
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
flower==2.0.1
gevent==24.11.1
geventhttpclient==2.3.1
gprof2dot==2024.6.6
greenlet==3.1.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
humanize==4.11.0
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.7.0
kombu==5.4.2
locust==2.32.2
MarkupSafe==3.0.2
model-bakery==1.20.0
msgpack==1.1.0
mysqlclient==2.2.5
oauthlib==3.2.2
openai==1.54.1
packaging==24.2
pillow==11.0.0
pipenv==2024.4.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus_client==0.21.0
prompt_toolkit==3.0.48
psutil==6.1.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.9.0
pytest==8.3.3
pytest-django==4.9.0
pytest-watch==4.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.2
pyzmq==26.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
tomli==2.1.0
tornado==6.4.1
tqdm==4.66.6
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
virtualenv==20.27.1
watchdog==6.0.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.8.2
zipp==3.21.0
zope.event==5.0
zope.interface==7.1.1
